+++ lib/cins_glass_effects.dart
@@ -11,6 +11,7 @@ export 'src/optics/glass_capabilities.dart';
 export 'src/optics/glass_render_path.dart';
 export 'src/optics/glass_filter_builder.dart';
 export 'src/optics/sdf_field.dart';
+export 'src/optics/sdf_texture_cache.dart';
 export 'src/materials/glass_material.dart';
 export 'src/materials/glass_presets.dart';
 export 'src/light/glass_light.dart';